Map and compare different UD smoothers for individual tracks

Map and compare different UD smoothers for individual tracks

Usage

opp_map_indUD(
  uds,
  ud_levels = c(50, 95),
  tracks,
  center = NULL,
  zoom = NULL,
  coast_scale = 50,
  viridis_option = "A"
)

Arguments

uds

Named list of estUDm or estUD objects to plot. As returned by opp_kernel.

ud_levels

List of numeric values between 1-99, indicting the utilization distribution levels to plot.

tracks

SpatialPointsDataFrame of location data used in opp_kernel.

center

Data frame containing columns 'Longitude' and 'Latitude' in decimal degrees, for plotting the colony or nest locations.

zoom

NULL or numeric value from 1:16, indicating the zoom level for map. If left as NULL (default) map extent will be defined by the bounding box of tracks. If numeric value is provided, bounding box will be cenetered on the mean location in tracks at the zoom level provided.

coast_scale

Mapping resolution for the coastline basemap. Must be one of: 10 - high resolution, 50 - medium resolution, 110 - low resolution.

viridis_option

A character string indicating the colormap option to use. Four options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D", the default option) and "cividis" (or "E").

Value

A named list of ggplot objects, with one slot for each individual in uds. #'@examples my_data <- opp_download_data(study = c(1247096889), login = NULL, start_month = NULL, end_month = NULL,season = NULL) my_track2kba <- opp2KBA(data = my_data) my_trips <- opp_get_trips(data = my_track2kba, innerBuff = 1, returnBuff = 10, duration = 1, gapLimit = 100, gapTime = 1, gapDist = 12, showPlots = F) my_interp <- ctcrw_interpolation(data = my_trips, site = my_track2kba$site, type = c('Complete','Incomplete','Gappy'), timestep = '10 min', interpolateGaps = F, showPlots = T, theta = c(8,2), quiet = F ) (hh <- opp_href(my_interp$interp)) my_ud_href1 <- opp_kernel(data = my_interp, interpolated = TRUE, smoother = hh/1, extendGrid = 30, res = 1) my_ud_href2 <- opp_kernel(data = my_interp, interpolated = TRUE, smoother = hh/2, extendGrid = 30, res = 1) my_plots_step <- opp_map_indUD( uds = list(Href1 = my_ud_href1,Href2 = my_ud_href2), tracks = my_interp$interp, center = my_track2kba$site, ud_levels = c(50,75,90), coast_scale = 50, viridis_option = 'A' ) my_plots_step[1] my_plots_step[2]