Generate reports for central place foraging species tracked in OPP

This function passes a table of parameters to an RMarkdown template to generate summary reports on tracking data analysis.

Usage

opp_reports_cpf(
  analysis_table = cpf_report_params,
  login = NULL,
  output_dir = "temp"
)

Arguments

analysis_table

Data frame that provides all values needed for the RMarkdown template, see details. Defaults to the cpf_report_params data object that is contained within OPPTools.

login

Stored Movebank login credentials if provided, otherwise function will prompt users to enter credentials.

output_dir

Name of of the directory where reports should be saved.

Details

Still need to write this up, once function is complete.

The example below will only work if the user has permissions for all the projects listed in cpf_report_params

Examples

print(cpf_report_params)
#>    mb_project_num                                  report_title
#> 1      1279263376         Cassin's auklet - Triangle Island, BC
#> 2      1912080185        Cassin's auklet - Frederick Island, BC
#> 3      1247096889           Tufted puffin - Triangle Island, BC
#> 4      1895716931          Ancient murrelet - George Island, BC
#> 5      1897273090          Ancient murrelet - Ramsay Island, BC
#> 6      1903294479          Rhinocerous auklet - Lucy Island, BC
#> 7      1904543040          Rhinocerous auklet - Pine Island, BC
#> 8      1904565280    Rhinocerous auklet - Protection Island, WA
#> 9      1903075528          Rhinocerous auklet - SGang Gwaay, BC
#> 10     1901463730      Rhinocerous auklet - Triangle Island, BC
#> 11     1953033045     Leach's storm-petrel - Copper Islands, BC
#> 12     1952967863     Leach's storm-petrel - Gillam Islands, BC
#> 13     1953016048       Leach's storm-petrel - Hippa Island, BC
#> 14     1948459965      Leach's storm-petrel - Storm Islands, BC
#> 15     1953094560   Fork-tailed storm-petrel - Hippa Island, BC
#> 16     1953062177 Fork-tailed storm-petrel - Copper Islands, BC
#> 17     1953086120 Fork-tailed storm-petrel - Gillam Islands, BC
#>                   file_name                      spp breeding_pairs minDist
#> 1    CAAU - Triangle Island          Cassin's auklet         550000     1.5
#> 2   CAAU - Frederick Island          Cassin's auklet             NA     1.5
#> 3    TUPU - Triangle Island            Tufted puffin             NA     1.0
#> 4      ANMU - George Island         Ancient murrelet          17000     1.5
#> 5      ANMU - Ramsay Island         Ancient murrelet          18000     1.0
#> 6        RHAU - Lucy Island       Rhinocerous auklet          25000     2.5
#> 7        RHAU - Pine Island       Rhinocerous auklet         140000     1.5
#> 8  RHAU - Protection Island       Rhinocerous auklet          34000     0.5
#> 9        RHAU - SGang Gwaay       Rhinocerous auklet             NA     2.5
#> 10   RHAU - Triangle Island       Rhinocerous auklet          42000     3.5
#> 11    LHSP - Copper Islands     Leach's storm-petrel             NA     5.0
#> 12    LHSP - Gillam Islands     Leach's storm-petrel             NA     5.0
#> 13      LHSP - Hippa Island     Leach's storm-petrel             NA     5.0
#> 14     LHSP - Storm Islands     Leach's storm-petrel             NA     5.0
#> 15      FTSP - Hippa Island Fork-tailed storm petrel             NA     5.0
#> 16    FTSP - Copper Islands Fork-tailed storm petrel             NA     5.0
#> 17    FTSP - Gillam Islands Fork-tailed storm petrel             NA     5.0
#>    maxDist minDur gapTime gapDist interpolateGaps timestep gridRes gridExtend
#> 1       20      1       1      12           FALSE    5 min       1         30
#> 2       10      1       1      12           FALSE    5 min       1         30
#> 3       10      1       1      18           FALSE    5 min       1         30
#> 4       20      1       3      19           FALSE   20 min       1         30
#> 5       20      1       3      19           FALSE   20 min       1         30
#> 6       20      1       1      12           FALSE    5 min       1         30
#> 7       10      1       1      12           FALSE    5 min       1         30
#> 8       10      1       1      12           FALSE    5 min       1         30
#> 9       10      1       1      12           FALSE    5 min       1         30
#> 10      10      1       1      12           FALSE    5 min       1         30
#> 11      60     12       6       5           FALSE   60 min       5        500
#> 12      60     12       6       5           FALSE   60 min       5        500
#> 13      60     12       6       5           FALSE   60 min       5        500
#> 14      60     12       6       5           FALSE   60 min       5        500
#> 15      60     12       6       5           FALSE   60 min       2        300
#> 16      60     12       6       5           FALSE   60 min       2        300
#> 17      60     12       6       5           FALSE   60 min       2        300
#>    speed mapZoom includeBB
#> 1     50     6.5         1
#> 2     50     6.5         1
#> 3     70     6.5         0
#> 4     75     6.5         1
#> 5     75     6.5         1
#> 6     50     6.5         1
#> 7     50     6.5         0
#> 8     50     6.5         1
#> 9     50     6.5         1
#> 10    50     6.5         1
#> 11    50     4.0         1
#> 12    50     4.5         1
#> 13    50     4.5         1
#> 14    50     4.0         0
#> 15    50     4.0         1
#> 16    50     6.0         0
#> 17    50     7.5         0
opp_reports_cpf()
#> username:
#> password:
#> Warning: path[1]="inst/rmarkdown/templates/opp-cpf-report/skeleton/skeleton.Rmd": No such file or directory
#> Error in abs_path(input): The file 'inst/rmarkdown/templates/opp-cpf-report/skeleton/skeleton.Rmd' does not exist.